refactor: use gateway package for cross-platform default gateway discovery
- Replace exec.Command with github.com/jackpal/gateway for pure Go implementation
- Simplify route_linux.go and route_bsd.go to use unified gateway discovery
- Fix IP_BOUND_IF issue on FreeBSD by removing socket option usage
- Linux supports Darwin, FreeBSD, and Linux
